requirements.txt
google-generativeai==0.8.3
streamlit>=1.30.0
pdfplumber>=0.10.0
openpyxl>=3.1.0
